refactor: output 객체 도메인과 분리
