+++ src/main/java/racinggame/controller/RacingGame.java
@@ -1,11 +1,11 @@
-package racinggame;
+package racinggame.controller;
 
 import racinggame.domain.Cars;
-import racinggame.domain.Referee;
 import racinggame.domain.Split;
 import racinggame.factory.CarFactory;
 import racinggame.strategy.RandomMove;
 import racinggame.ui.InputView;
+import racinggame.ui.OutputView;
 
 import java.util.Scanner;
 
@@ -17,9 +17,12 @@ public class RacingGame {
         int racingCount = inputView.racingCount();
 
         Cars cars = new Cars(new CarFactory().generateCar(new Split().toCarNames(carName), new RandomMove()));
-        Referee referee = new Referee();
-        referee.startRacing(racingCount, cars);
-        referee.judgeWinners(cars);
+        OutputView outputView = new OutputView();
+
+        for (int step = 0; step < racingCount; step++) {
+            outputView.printRacing(cars.racing());
+        }
+        outputView.printWinners(cars.findWinners());
     }
 
 }
